More file manipulation.
The column removal tool will enumerate through a large number of
workbooks and eliminate one or more predefined columns. The
get_files_in_directory code will list every file name in a directory,
which can be useful when mass-processing files; it also illustrates use
of the Application.FileDialog method with msoFileDialogFolderPicker as
an option.